3.25会议实验代码/requirements（对dso，python版本为3.7）.txt
baostock==0.8.8
gymnasium==0.28.1
matplotlib==3.3.4
numpy==1.21.0
pandas==1.2.4
pyqlib
sb3_contrib==2.0.0
stable_baselines3==2.0.0
torch==1.13.1
tensorflow==1.14.0
shimmy==1.1.0
fire
Flask==2.2.5
tqdm
tensorboard==1.14.0